backend/requirements.txt
flask
flask-cors
torch
torchvision
transformers
datasets
pandas
pillow
safetensors
scikit-learn
numpy 